ui(refactor): split 'superdeks-ingest' module.
