+++ scripts/superdesk-ingest/controllers/index.js
@@ -1,0 +1,4 @@
+export { IngestListController } from './IngestListController';
+export { IngestSettingsController } from './IngestSettingsController';
+export { IngestDashboardController } from './IngestDashboardController';
+export { ExternalSourceController } from './ExternalSourceController';
