+++ scripts/superdesk-ingest/directives/IngestRoutingAction.js
@@ -1,0 +1,95 @@
+IngestRoutingAction.$inject = ['desks', 'macros', 'subscribersService', 'metadata', 'gettext'];
+export function IngestRoutingAction(desks, macros, subscribersService, metadata, gettext) {
+    return {
+        scope: {rule: '='},
+        templateUrl: 'scripts/superdesk-ingest/views/settings/ingest-routing-action.html',
+        link: function(scope) {
+            scope.newFetch = {};
+            scope.newPublish = {};
+            scope.deskLookup = {};
+            scope.stageLookup = {};
+            scope.macroLookup = {};
+            scope.customSubscribers = [];
+            scope.target_types = [];
+
+            desks.initialize()
+            .then(function() {
+                scope.deskLookup = desks.deskLookup;
+                scope.stageLookup = desks.stageLookup;
+            });
+
+            macros.get(true).then(function(macros) {
+                _.transform(macros, function(lookup, macro, idx) {
+                    scope.macroLookup[macro.name] = macro;
+                });
+            });
+
+            subscribersService.fetchActiveSubscribers().then(function(items) {
+                scope.customSubscribers = [];
+                _.each(items, function(item) {
+                    scope.customSubscribers.push({'_id': item._id, 'name': item.name});
+                });
+            });
+
+            metadata.initialize()
+                .then(function() {
+                    scope.target_types = metadata.values.subscriberTypes;
+                });
+
+            scope.getActionString = function(action) {
+                if (scope.deskLookup[action.desk] && scope.stageLookup[action.stage]) {
+                    var actionValues = [];
+                    actionValues.push(scope.deskLookup[action.desk].name);
+                    actionValues.push(scope.stageLookup[action.stage].name);
+                    if (action.macro) {
+                        actionValues.push(scope.macroLookup[action.macro].label);
+                    } else {
+                        actionValues.push(' - ');
+                    }
+                    if (action.target_subscribers && action.target_subscribers.length > 0) {
+                        actionValues.push(_.map(action.target_subscribers, 'name').join(','));
+                    } else {
+                        actionValues.push(' - ');
+                    }
+                    if (action.target_types && action.target_types.length > 0) {
+                        var targets = [];
+                        _.forEach(action.target_types, function(target_type) {
+                            targets.push((!target_type.deny ? gettext('Not ') : '') + target_type.name);
+                        });
+                        actionValues.push(targets.join(','));
+                    }
+
+                    return actionValues.join(' / ');
+                }
+            };
+
+            scope.addFetch = function() {
+                if (scope.newFetch.desk && scope.newFetch.stage) {
+                    scope.rule.actions.fetch.push(scope.newFetch);
+                    scope.newFetch = {};
+                }
+            };
+
+            scope.removeFetch = function(fetchAction) {
+                _.remove(scope.rule.actions.fetch, function(f) {
+                    return f === fetchAction;
+                });
+            };
+
+            scope.addPublish = function() {
+                if (scope.newPublish.desk && scope.newPublish.stage) {
+                    scope.rule.actions.publish.push(scope.newPublish);
+                    scope.newPublish = {};
+                    scope.newPublish.target_subscribers = [];
+                    scope.newPublish.target_types = [];
+                }
+            };
+
+            scope.removePublish = function(publishAction) {
+                _.remove(scope.rule.actions.publish, function(p) {
+                    return p === publishAction;
+                });
+            };
+        }
+    };
+}
