+++ scripts/superdesk-ingest/directives/IngestUserDashboardList.js
@@ -1,0 +1,10 @@
+IngestUserDashboardList.$inject = [];
+export function IngestUserDashboardList () {
+    return {
+        templateUrl: 'scripts/superdesk-ingest/views/dashboard/ingest-dashboard-widget-list.html',
+        scope: {
+            items: '=',
+            setUserPreferences: '&'
+        }
+    };
+}
