+++ scripts/superdesk-ingest/directives/index.js
@@ -1,0 +1,12 @@
+export { PieChartDashboard } from './PieChartDashboard';
+export { IngestSourcesContent } from './IngestSourcesContent';
+export { IngestRulesContent } from './IngestRulesContent';
+export { IngestRoutingContent } from './IngestRoutingContent';
+export { IngestRoutingGeneral } from './IngestRoutingGeneral';
+export { IngestRoutingAction } from './IngestRoutingAction';
+export { IngestRoutingSchedule } from './IngestRoutingSchedule';
+export { IngestRoutingFilter } from './IngestRoutingFilter';
+export { SortRules } from './SortRules';
+export { IngestUserDashboard } from './IngestUserDashboard';
+export { IngestUserDashboardList } from './IngestUserDashboardList';
+export { IngestUserDashboardDropdown } from './IngestUserDashboardDropdown';
