+++ scripts/superdesk-ingest/services/index.js
@@ -1,0 +1,4 @@
+export { IngestProviderService } from './IngestProviderService';
+export { SubjectService } from './SubjectService';
+export { RemoveIngestedService } from './RemoveIngestedService';
+export { SendService } from './SendService';
