+++ scripts/superdesk-ingest/tests/ingest.spec.js
